# supervisor-console/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zippy Logitech API",
    "description": "Logistics Business Intelligence Platform API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/dev-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Dev Login",
        "description": "Create/update a local dev user and return a JWT.\n\nThis intentionally avoids OAuth while keeping browser tests and local\noperator workflows on the same Authorization header path as production.",
        "operationId": "dev_login_api_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/{order_id}/bids": {
      "post": {
        "tags": [
          "Bidding"
        ],
        "summary": "Create Bid",
        "description": "Place a bid on an order",
        "operationId": "create_bid_api_v1_orders__order_id__bids_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Bidding"
        ],
        "summary": "List Bids",
        "description": "List all bids for an order",
        "operationId": "list_bids_api_v1_orders__order_id__bids_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by bid status",
              "title": "Status"
            },
            "description": "Filter by bid status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bids/{bid_id}/accept": {
      "post": {
        "tags": [
          "Bidding"
        ],
        "summary": "Accept Bid",
        "description": "Accept a bid",
        "operationId": "accept_bid_api_v1_bids__bid_id__accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bid_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bid Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bids/{bid_id}/reject": {
      "post": {
        "tags": [
          "Bidding"
        ],
        "summary": "Reject Bid",
        "description": "Reject a bid",
        "operationId": "reject_bid_api_v1_bids__bid_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bid_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bid Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bids/{bid_id}/counter": {
      "post": {
        "tags": [
          "Bidding"
        ],
        "summary": "Counter Bid",
        "description": "Counter-offer on a bid",
        "operationId": "counter_bid_api_v1_bids__bid_id__counter_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bid_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bid Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidCounter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BidCounter": {
        "properties": {
          "counter_amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Counter Amount"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "counter_amount"
        ],
        "title": "BidCounter"
      },
      "BidCreate": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vehicle Id"
          },
          "driver_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Driver Name"
          },
          "driver_phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Driver Phone"
          },
          "bid_amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Bid Amount"
          },
          "estimated_eta_hours": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Eta Hours"
          },
          "estimated_arrival_hours": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Arrival Hours"
          },
          "vehicle_available_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Available At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "driver_name",
          "driver_phone",
          "bid_amount"
        ],
        "title": "BidCreate"
      },
      "BidListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "bids": {
            "items": {
              "$ref": "#/components/schemas/BidResponse"
            },
            "type": "array",
            "title": "Bids"
          }
        },
        "type": "object",
        "required": [
          "total",
          "bids"
        ],
        "title": "BidListResponse"
      },
      "BidResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "vehicle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vehicle Id"
          },
          "driver_name": {
            "type": "string",
            "title": "Driver Name"
          },
          "driver_phone": {
            "type": "string",
            "title": "Driver Phone"
          },
          "bid_amount": {
            "type": "number",
            "title": "Bid Amount"
          },
          "counter_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Amount"
          },
          "estimated_eta_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Eta Hours"
          },
          "estimated_arrival_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Arrival Hours"
          },
          "vehicle_available_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Available At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "vehicle_id",
          "driver_name",
          "driver_phone",
          "bid_amount",
          "counter_amount",
          "estimated_eta_hours",
          "estimated_arrival_hours",
          "vehicle_available_at",
          "notes",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "BidResponse"
      },
...
      "DevLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Password"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ],
            "default": "ops_admin"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "DevLoginRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role",
          "username"
        ],
        "title": "TokenResponse"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "customer",
          "driver",
          "transport_company",
          "supervisor",
          "support_admin",
          "ops_admin",
          "finance_admin",
          "super_admin"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}